Clean up CrStackFrame
